change names and refactor few more functions
