+++ nums/numpy/api/equality.py
