+++ nums/numpy/api/reduction.py
@@ -18,7 +18,7 @@
 from nums.core.application_manager import instance as _instance
 from nums.core.array.blockarray import BlockArray
 
-from nums.numpy.api.utility import *
+from nums.numpy.api.equality import *
 
 ############################################
 # Reduction Ops
