+++ nums/numpy/api/special.py
@@ -1,0 +1,52 @@
+# coding=utf-8
+# Copyright (C) 2020 NumS Development Team.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# pylint: disable = redefined-builtin, too-many-lines, anomalous-backslash-in-string, unused-wildcard-import, wildcard-import
+from typing import Tuple
+
+from nums.core.application_manager import instance as _instance
+from nums.core.array.blockarray import BlockArray
+
+from nums.numpy.api.reduction import *
+from nums.numpy.api.generated import *
+
+
+############################################
+# Special
+############################################
+
+
+def top_k(
+    a: BlockArray, k: int, largest=True, sorted=False
+) -> Tuple[BlockArray, BlockArray]:
+    """Find the `k` largest or smallest elements of a BlockArray.
+
+    If there are multiple kth elements that are equal in value, then no guarantees are made as
+    to which ones are included in the top k.
+
+    Args:
+        a: A BlockArray.
+        k: Number of top elements to return.
+        largest: Whether to return largest or smallest elements.
+
+    Returns:
+        A tuple containing two BlockArrays, (`values`, `indices`).
+        values: Values of the top k elements, unsorted.
+        indices: Indices of the top k elements, ordered by their corresponding values.
+    """
+    if sorted:
+        # The result can be sorted when sorting is implemented.
+        raise NotImplementedError("'sorted' is currently not supported.")
+    return _instance().top_k(a, k, largest=largest)
